Feed animal fix + Mining code change to update weapon/use existing actions
The code used to mine minerals/cut trees was adjusted to no longer kill animals in the interact menu. Also I changed it so that it only displayed the pickaxe/axe when they are selected. Also I altered them to use the existing input actions. PrimaryClick & Interact instead of the previous two actions.